Âupdated userScoreAgg to LeaderboardAggResult & relocate packg
